ES-1638

+++ oidc-ui/configure_start.sh
@@ -5,44 +5,48 @@ set -e
 
 echo "Downloading pre-requisites started."
 
-#i18n bundle
-echo "Downloading i18n bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/i18n/esignet-i18n-bundle.zip -O $i18n_path/esignet-i18n-bundle.zip
+# Check if $i18n_url_env is not empty
+if [[ -n "$i18n_url_env" ]]; then
+    echo "i18n_url_env is set: $i18n_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $i18n_url_env -O $i18n_path/esignet-i18n-bundle.zip
 
-echo "unzip i18n bundle files.."
-chmod 775 $i18n_path/*
+    echo "unzip i18n bundle files.."
+    chmod 775 $i18n_path/*
+    cd $i18n_path
+    unzip -o esignet-i18n-bundle.zip
+    rm esignet-i18n-bundle.zip
+    echo "unzip i18n bundle completed."
+fi
 
-cd $i18n_path
-unzip -o esignet-i18n-bundle.zip
-rm esignet-i18n-bundle.zip
+# Check if $theme_url_env is not empty
+if [[ -n "$theme_url_env" ]]; then
+    echo "theme_url_env is set: $theme_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $theme_url_env -O $theme_path/esignet-theme.zip
 
-#theme bundle
-echo "Downloading theme bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/theme/esignet-theme.zip -O $theme_path/esignet-theme.zip
+    echo "unzip theme files.."
+    chmod 775 $theme_path/*
+    cd $theme_path
+    unzip -o esignet-theme.zip
+    rm esignet-theme.zip
+    echo "unzip theme completed."
+fi
 
-echo "unzip i18n bundle files.."
-chmod 775 $theme_path/*
+# Check if $images_url_env is not empty
+if [[ -n "$images_url_env" ]]; then
+    echo "images_url_env is set: $images_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $images_url_env -O $image_path/esignet-image.zip
 
-cd $theme_path
-unzip -o esignet-theme.zip
-rm esignet-theme.zip
-
-#image bundle
-echo "Downloading image bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/image/esignet-image.zip -O $image_path/esignet-image.zip
-
-echo "unzip image bundle files.."
-chmod 775 $image_path/*
-
-cd $image_path
-unzip -o esignet-image.zip
-rm esignet-image.zip
+    echo "unzip image files.."
+    chmod 775 $image_path/*
+    cd $image_path
+    unzip -o esignet-image.zip
+    rm esignet-image.zip
+    echo "unzip image completed."
+fi
 
 #sign-in-button-plugin
 echo "Downloading plugins"
-
 wget --no-check-certificate --no-cache --no-cookies $SIGN_IN_WITH_ESIGNET_PLUGIN_URL -O $plugins_path/temp/sign-in-button-plugin.zip
-
 echo "unzip plugins.."
 cd $plugins_path/temp
 unzip -o sign-in-button-plugin.zip
